Make all tracks draggable!
